oops/requirements.txt
aliyun-log-python-sdk==0.6.23
aliyun-python-sdk-cdn==2.6.2
aliyun-python-sdk-cms==6.0.6
aliyun-python-sdk-core-v3==2.5.2
aliyun-python-sdk-ecs==4.4.3
aliyun-python-sdk-slb==3.2.4
amqp==2.3.2
ansi2html==1.3.0
ansible==2.4.2.0
anyjson==0.3.3
asn1crypto==0.23.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
cached-property==1.3.1
celery==4.2.1
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
decorator==4.1.2
Django==1.11.7
django-auth-ldap==1.3.0
django-celery-beat==1.3.0
django-celery-results==1.0.4
django-cors-headers==2.4.0
django-crontab==0.7.1
django-filter==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
django-timezone-field==3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
dnspython==1.15.0
enum34==1.1.6
exchangelib==1.10.7
flower==0.9.2
future==0.16.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.10
kombu==4.2.1
lxml==4.1.1
MarkupSafe==1.0
netifaces==0.10.7
ntlm-auth==1.0.6
paramiko==2.4.0
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pika==0.12.0
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.6
ptyprocess==0.5.2
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.6.4
pyldap==2.4.45
PyMySQL==0.7.11
PyNaCl==1.2.1
python-crontab==2.3.5
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzabbix==0.7.4
redis==2.10.6
requests==2.18.4
requests-ntlm==1.1.0
setproctitle==1.1.10
shadowsocks==2.8.2
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
sqlparse==0.2.4
tornado==5.1.1
traitlets==4.3.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
